separate display notification for easy re-use
