+++ UI/pages/article.html
@@ -113,6 +113,7 @@
      https://firebase.google.com/docs/web/setup#available-libraries -->
     <script src="https://www.gstatic.com/firebasejs/7.17.1/firebase-analytics.js"></script>
     <script src="../scripts/firebase.config.js" type="module"></script>
+    <script src="../scripts/helperFunctions.js" type="module"></script>
     <script src="../scripts/article.js" type="module"></script>
   </body>
 </html>
