+++ UI/pages/queries.html
@@ -119,6 +119,7 @@
         &copy; Elvis Iraguha 2020
       </footer>
     </div>
+    <script src="../scripts/helperFunctions.js" type="module"></script>
     <script src="../scripts/queriesList.js" type="module"></script>
     <script src="../scripts/queries.js" type="module"></script>
   </body>
