+++ UI/pages/update-profile.html
@@ -220,6 +220,7 @@
         &copy; Elvis Iraguha 2020
       </footer>
     </div>
+    <script src="../scripts/helperFunctions.js" type="module"></script>
     <script src="../scripts/profileData.js" type="module"></script>
     <script src="../scripts/update-profile.js" type="module"></script>
   </body>
