+++ UI/pages/write.html
@@ -82,6 +82,7 @@
         &copy; Elvis Iraguha 2020
       </footer>
     </div>
-    <script src="../scripts/write.js"></script>
+    <script src="../scripts/helperFunctions.js" type="module"></script>
+    <script src="../scripts/write.js" type="module"></script>
   </body>
 </html>
